requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.9
Jinja2==3.1.4
slowapi==0.1.9
email-validator==2.2.0
